Added Integration Test for Job
Added Integration Test for Job Create (to fill the screen with data) but both chrome and safari (allow remote automation) browsers are unable to open the url. It opens a new instance of the browser and cant connect to the server. Tried to call exe file from the bin/debug/net5 folder using command line but file not found.